move TypeScriptZod to own dir
