+++ packages/quicktype-core/src/language/TypeScriptZod/index.ts
@@ -1,0 +1,2 @@
+export { TypeScriptZodTargetLanguage, typeScriptZodOptions } from "./language";
+export { TypeScriptZodRenderer } from "./TypeScriptZodRenderer";
